Почти готово Доделать: -Видео -Отправка на почту в конце дня -ReadMe -Gitignore -Doc

+++ src/main/java/com/mrokos/RelexProject/dtos/UserResponseDto.java
@@ -4,11 +4,11 @@ import lombok.Data;
 
 @Data
 public class UserResponseDto {
-    private Integer id;
+    private Long id;
     private String username;
     private String email;
 
-    public UserResponseDto(Integer id, String username, String email) {
+    public UserResponseDto(Long id, String username, String email) {
         this.id = id;
         this.username = username;
         this.email = email;
